remove lower
