+++ python/tvm/__init__.py
@@ -55,7 +55,7 @@ from . import target
 from . import te
 
 # tvm.driver
-from .driver import build, lower
+from .driver import build
 
 # others
 from . import arith
