+++ python/tvm/tir/__init__.py
@@ -108,3 +108,4 @@ from . import ir_builder
 from . import transform
 from . import analysis
 from . import stmt_functor
+from .build import build
