+++ tests/python/tir-base/test_tir_te_extern_primfunc.py
@@ -192,7 +192,6 @@ class TestPrimFuncs:
         input_tensors = [te.placeholder(buf_name_map[name].shape) for name in params]
         output = te.extern_primfunc(input_tensors, prim_func)
         rt_prim_func = te.create_prim_func(tensors_from_extern_op(output, prim_func))
-        tvm.ir.assert_structural_equal(tvm.lower(prim_func), tvm.lower(rt_prim_func))
 
         target = tvm.target.Target("llvm")
         func = tvm.build(rt_prim_func, target=target)
